+++ packages/Input/src/Input.tsx
@@ -28,11 +28,13 @@ export interface InputProps extends React.ComponentPropsWithRef<'input'> {
   // True if you need the input to be focus on page load.
   autoFocus?: boolean;
   // True if you need the input to be readonly.
-  readOnly?: boolean;
+  disabled?: boolean;
   // Function called when the value changes.
   onChange?: (e: any) => void;
   // Add a data-test tag for automated tests
   dataTest?: string;
+  // Use a prefix for add a icon before the input text
+  prefixIcon?: React.ReactNode;
 }
 
 const Input: React.FunctionComponent<InputProps> = React.forwardRef(
@@ -44,16 +46,18 @@ const Input: React.FunctionComponent<InputProps> = React.forwardRef(
       isCompact,
       value,
       autoFocus,
-      readOnly,
+      disabled,
       onChange,
       dataTest,
+      prefixIcon,
       ...rest
     } = props;
 
     const classes = cx('ids-input', className, {
       'ids-input--compact': isCompact,
       'ids-input--error': error,
-      'ids-input--readonly': readOnly,
+      'ids-input--disabled': disabled,
+      'ids-input--prefixIcon': prefixIcon,
     });
 
     useEffect(() => {
@@ -101,17 +105,32 @@ const Input: React.FunctionComponent<InputProps> = React.forwardRef(
       }
     };
 
-    return (
+    const inputRender = (
       <input
         ref={ref}
         className={classes}
         type={type}
-        value={value}
-        readOnly={readOnly}
+        defaultValue={value}
+        readOnly={disabled}
         onChange={handleOnChange}
         data-test={dataTest}
         {...rest}
       />
+    );
+
+    return prefixIcon ? (
+      <div className="ids-input__wrapper">
+        <span
+          className={cx('ids-input__asset', {
+            'ids-input__asset--compact': isCompact,
+          })}
+        >
+          {prefixIcon}
+        </span>
+        {inputRender}
+      </div>
+    ) : (
+      inputRender
     );
   }
 );