+++ typings/icons.d.ts
@@ -23,3 +23,21 @@ declare module '@igloo-ui/icons/dist/Plus' {
     size?: 'small' | 'medium' | 'large';
   }): JSX.Element;
 }
+
+declare module '@igloo-ui/icons/dist/Settings' {
+  export default function Start(props: {
+    width?: number;
+    height?: number;
+    fill?: string;
+    size?: 'small' | 'medium' | 'large';
+  }): JSX.Element;
+}
+
+declare module '@igloo-ui/icons/dist/Search' {
+  export default function Start(props: {
+    width?: number;
+    height?: number;
+    fill?: string;
+    size?: 'small' | 'medium' | 'large';
+  }): JSX.Element;
+}