+++ website/components/Tag.js
@@ -1,0 +1,13 @@
+import cx from 'classnames';
+
+export default function Tag({ children, className, appearance }) {
+  return (
+    <span
+      className={cx('io-tag', className, {
+        [`io-tag--${appearance}`]: appearance,
+      })}
+    >
+      {children}
+    </span>
+  );
+}