+++ website/components/icons/index.js
@@ -1,0 +1,2 @@
+export { default as Copy } from './Copy';
+export { default as Check } from './Check';