+++ packages/Input/CHANGELOG.md
@@ -1,14 +1,31 @@
 # @igloo-ui/input
 
+## 1.0.0
+
+### Major Changes
+
+- ef50a1b: BRAKING CHANGE
+
+  - Replace `readOnly` props by `disabled`
+
+  New Features
+
+  - Add options for passing a Icon on the input
+
+  Update
+
+  - Upgrade tokens version
+
 ## 0.1.0
+
 ### Minor Changes
 
 - 8e6a4ec: ## Input component
-  
+
   The Input component specifies an input field where the user can enter data.
-  
+
   ### Component API
-  
+
   | name      | description                                          |
   | --------- | ---------------------------------------------------- |
   | type      | Specifies the type to display                        |